save-to-text/requirements.txt
google-cloud-pubsub==2.5.0
google-cloud-storage==1.38.0